IntelliJ Idea 12 Eclipse keymap for OS X
The Eclipse keymap for IntelliJ Idea 12 on OS X doesn't follow the OS X
convention of using the Command key instead of the Ctrl key. This is a
customized version of the Eclipse keymap that replaces some of the
keyboard shortcuts that I use the most from Eclipse so that they use
the Command key instead.